Move initialization code from main.go to the registry package
This makes it easier to embed a registry instance inside another
application.
